requirements.txt
discord~=1.0.1
mysql-connector-python~=8.0.26
selenium~=3.141.0